Created my first OOP program
In my first OOP program I have created a simulation email service. It's features include an inbox with pre-written emails, an option to import a new email, view unread emails, spam folder and send which once updated will send to an outbox.

Emails can also be moved to the spam folder or deleted.